refactor: 의ì존성 분리 및 중복 ã제거, 코드 모듈화, 코드리뷰 수정
